// packages/tsts/src/internal/ast/generated/data.ts
export interface ImportDeclaration extends StatementBase, ModifiersBase, CompositeBase, DeclarationBase {
  ImportClause: GoPtr<ImportClauseNode>;
  ModuleSpecifier: GoPtr<Expression>;
  Attributes: GoPtr<ImportAttributesNode>;
}
// ...
export interface NamespaceImport extends NodeBase, DeclarationBase, ExportableBase {
  name: GoPtr<IdentifierNode>;
}
// ...
export interface NamedImports extends NodeBase, CompositeBase {
  Elements: GoPtr<ImportSpecifierList>;
}
// ...
export interface ImportClause extends NodeBase, DeclarationBase, ExportableBase, CompositeBase {
  PhaseModifier: Kind;
  name: GoPtr<IdentifierNode>;
  NamedBindings: GoPtr<NamedImportBindings>;
}